+++ include/libtorrent/kademlia/msg.hpp
@@ -30,16 +30,17 @@ POSSIBILITY OF SUCH DAMAGE.
 
 */
 
-#ifndef MSG_HPP
-#define MSG_HPP
+#ifndef TORRENT_KADEMLIA_MSG_HPP
+#define TORRENT_KADEMLIA_MSG_HPP
 
 #include <string>
-#include <libtorrent/kademlia/node_id.hpp>
-#include <boost/asio/ip/udp.hpp>
+#include "libtorrent/socket.hpp"
+#include "libtorrent/kademlia/node_id.hpp"
 
 namespace libtorrent {
 
 struct bdecode_node;
+struct entry;
 
 namespace dht {
 
@@ -60,6 +61,40 @@ private:
 	// explicitly disallow assignment, to silence msvc warning
 	msg& operator=(msg const&);
 };
+
+struct key_desc_t
+{
+	char const* name;
+	int type;
+	int size;
+	int flags;
+
+	enum {
+		// this argument is optional, parsing will not
+		// fail if it's not present
+		optional = 1,
+		// for dictionaries, the following entries refer
+		// to child nodes to this node, up until and including
+		// the next item that has the last_child flag set.
+		// these flags are nestable
+		parse_children = 2,
+		// this is the last item in a child dictionary
+		last_child = 4,
+		// the size argument refers to that the size
+		// has to be divisible by the number, instead
+		// of having that exact size
+		size_divisible = 8
+	};
+};
+
+// verifies that a message has all the required
+// entries and returns them in ret
+bool TORRENT_EXPORT verify_message(bdecode_node const& msg, key_desc_t const desc[]
+	, bdecode_node ret[], int size, char* error, int error_size);
+
+// generate an error response message
+void TORRENT_EXPORT incoming_error(entry& e, char const* msg, int error_code = 203);
+
 
 } }
